+++ index.js
@@ -5,6 +5,15 @@ const Base64 = require("js-base64").Base64;
 const YAML = require("yaml");
 const process = require("process");
 
+const {
+  getContents,
+  getHeadSha,
+  getSha,
+  getLatestImageUrl
+} = require("./utils");
+
+const { closePRs, createPR } = require("./prUtils")
+
 // Environmment ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 const myToken = process.env.TOKEN;
@@ -55,204 +64,6 @@ const PROJECTS = [
   },
 ];
 
-// Logic ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-const getCommitMessages = async (repo, sha) => {
-  const { data: commits } = await octokit.repos.listCommits({
-    owner: GH_CDS,
-    repo,
-    per_page: 50,
-  });
-  let index = 0;
-  for (let i = 0; i < 50; i++) {
-    if (commits[i].sha.startsWith(sha)) {
-      index = i;
-      break;
-    }
-  }
-  return commits
-    .slice(0, index)
-    .map(
-      (c) =>
-        `- [${c.commit.message.split("\n\n")[0]}](${c.html_url}) by ${c.commit.author.name
-        }`
-    );
-};
-
-const getHeadSha = async (repo) => {
-  const { data: repoDetails } = await octokit.repos.get({
-    owner: GH_CDS,
-    repo,
-  });
-  const { data: repoBranch } = await octokit.repos.getBranch({
-    owner: GH_CDS,
-    repo,
-    branch: repoDetails.default_branch,
-  });
-  return repoBranch.commit.sha;
-};
-
-const getLatestTag = async (repo) => {
-  const {
-    data: [latestTag],
-  } = await octokit.repos.listTags({
-    owner: GH_CDS,
-    repo,
-    per_page: 1,
-  });
-
-  return latestTag.name;
-};
-
-// Close old auto PRs
-async function closePRs() {
-  const { data: prs } = await octokit.pulls.list({
-    owner: GH_CDS,
-    repo: "notification-manifests",
-    state: "open",
-  });
-
-  prs.forEach(async (pr) => {
-    if (pr.title.startsWith("SJA TEST [AUTO-PR]")) {
-      await octokit.pulls.update({
-        owner: GH_CDS,
-        repo: "notification-manifests",
-        pull_number: pr.number,
-        state: "closed",
-      });
-      await octokit.git.deleteRef({
-        owner: GH_CDS,
-        repo: "notification-manifests",
-        ref: `heads/${pr.head.ref}`,
-      });
-    }
-  });
-}
-
-async function isNotLatestManifestsVersion() {
-  const releaseConfig = await getContents(
-    GH_CDS,
-    "notification-manifests",
-    "env/production/kustomization.yaml"
-  );
-
-  const releaseContent = Base64.decode(releaseConfig.content);
-  const prodVersion = releaseContent.match(
-    /notification-manifests\/\/base\?ref=(.*)/
-  )[1];
-
-  const latestVersion = await getLatestTag("notification-manifests");
-
-  return prodVersion != latestVersion;
-}
-
-async function isNotLatestTerraformVersion() {
-  const prodWorkflow = await getContents(
-    GH_CDS,
-    "notification-terraform",
-    ".github/workflows/infrastructure_version.txt"
-  );
-
-  const prodVersion = Base64.decode(prodWorkflow.content).trim();
-  const latestVersion = (await getLatestTag("notification-terraform")).replace(
-    "v",
-    ""
-  );
-
-  return prodVersion != latestVersion;
-}
-
-function shortSha(fullSha) {
-  return fullSha.slice(0, 7);
-}
-
-function getSha(imageName) {
-  return imageName.split(":").slice(-1)[0];
-}
-
-function getLatestImageUrl(projectName, headSha) {
-  const ecrUrl = PROJECTS.filter(project => project["ecrName"] == projectName)[0].ecrUrl
-  return `${ecrUrl}/${projectName}:${shortSha(headSha)}`;
-}
-
-async function buildLogs(projects) {
-  let logs = projects.map(async (project) => {
-    const msgsCommits = await getCommitMessages(project.name, project.oldSha);
-    const strCommits = msgsCommits.join("\n");
-    const projectName = project.name.toUpperCase();
-    return `${projectName}\n\n${strCommits}`;
-  });
-  logs = await Promise.all(logs);
-
-  logs = logs.join("\n\n");
-  if (await isNotLatestManifestsVersion()) {
-    logs = `⚠️ **The production version of manifests is behind the latest staging version. Consider upgrading to the latest version before merging this pull request.** \n\n ${logs}`;
-  }
-
-  if (await isNotLatestTerraformVersion()) {
-    logs = `⚠️ **The production version of the Terraform infrastructure is behind the latest staging version. Consider upgrading to the latest version before merging this pull request.** \n\n ${logs}`;
-  }
-  return logs;
-}
-
-async function getContents(owner, repo, path) {
-  const { data: data } = await octokit.repos.getContents({
-    owner,
-    repo,
-    path,
-  });
-  return data;
-}
-
-async function createPR(
-  projects,
-  issueContent,
-  releaseContentArray
-) {
-  const branchName = `release-${new Date().getTime()}`;
-  const manifestsSha = await getHeadSha("notification-manifests");
-  const logs = await buildLogs(projects);
-
-  const ref = await octokit.git.createRef({
-    owner: GH_CDS,
-    repo: "notification-manifests",
-    ref: `refs/heads/${branchName}`,
-    sha: manifestsSha,
-  });
-
-  const manifestUpdates = projects
-    .map((project) => {
-      return `${project.name}:${shortSha(project.headSha)}`;
-    })
-    .join(" and ");
-
-  const updates = releaseContentArray.map(async ({ manifestFile, releaseContent, newReleaseContentBlob }) => {
-    return await octokit.repos.createOrUpdateFile({
-      owner: GH_CDS,
-      repo: "notification-manifests",
-      branch: branchName,
-      sha: releaseContent.sha,
-      path: manifestFile,
-      message: `Updated manifests to ${manifestUpdates}`,
-      content: newReleaseContentBlob,
-    })
-  })
-  await Promise.all(updates)
-
-  const pr = await octokit.pulls.create({
-    owner: GH_CDS,
-    repo: "notification-manifests",
-    title: `SJA TEST [AUTO-PR] Automatically generated new release ${new Date().toISOString()}`,
-    head: branchName,
-    base: "main",
-    body: issueContent.replace(
-      "> Give details ex. Security patching, content update, more API pods etc",
-      logs
-    ),
-    draft: true,
-  });
-  return Promise.all([ref, pr] + updates);
-}
 
 async function hydrateWithSHAs(releaseConfig, projects) {
   return await Promise.all(
@@ -265,6 +76,7 @@ async function hydrateWithSHAs(releaseConfig, projects) {
       }
       matchingProject.headSha = await getHeadSha(matchingProject.name);
       matchingProject.headUrl = getLatestImageUrl(
+        PROJECTS,
         matchingProject.ecrName,
         matchingProject.headSha
       );
@@ -314,8 +126,10 @@ async function run() {
     return;
   }
 
-  await closePRs();
-  await createPR(PROJECTS, issueContent, changesToManifestFiles);
+  await closePRs(GH_CDS);
+  await createPR(GH_CDS, PROJECTS, issueContent, changesToManifestFiles);
+
+  console.log("done")
 }
 
 // Main execute ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
